+++ src/app/[lang]/blog/[slug]/page.tsx
@@ -2,6 +2,8 @@ import * as path from 'node:path';
 
 import PostIndexPage from 'src/components/blog/postIndexPage';
 
+import { Metadata } from 'next';
+
 import { Language } from 'src/types/language';
 import { getPostsNames } from 'src/services/blog';
 
@@ -20,6 +22,22 @@ type Params = {
     slug: string;
   };
 };
+
+export async function generateMetadata({ params }: Params): Promise<Metadata> {
+  const baseUrl = 'https://kavoon.com.ua';
+  const { lang, slug } = params;
+
+  return {
+    alternates: {
+      canonical: `${baseUrl}/${lang}/blog/${slug}`,
+      languages: {
+        en: `${baseUrl}/en/blog/${slug}`,
+        uk: `${baseUrl}/uk/blog/${slug}`,
+        pl: `${baseUrl}/pl/blog/${slug}`,
+      },
+    },
+  };
+}
 
 async function PostPage({ params: { lang, slug } }: Params) {
   return <PostIndexPage language={lang} slug={slug} postspath={postspath} />;