+++ src/app/[lang]/blog/page.tsx
@@ -1,4 +1,5 @@
 import * as path from 'node:path';
+import { Metadata } from 'next';
 
 import BlogIndexPage from 'src/components/blog/blogIndexPage';
 
@@ -8,6 +9,23 @@ import { getPosts } from 'src/services/blog';
 type Props = {
   params: { lang: Language };
 };
+
+export function generateMetadata({ params }: Props): Metadata {
+  const baseUrl = 'https://kavoon.com.ua';
+  const { lang } = params;
+  const path = '/blog';
+
+  return {
+    alternates: {
+      canonical: `${baseUrl}/${lang}${path}`,
+      languages: {
+        en: `${baseUrl}/en${path}`,
+        uk: `${baseUrl}/uk${path}`,
+        pl: `${baseUrl}/pl${path}`,
+      },
+    },
+  };
+}
 
 async function Blog({ params: { lang } }: Props) {
   const postspath = path.join(process.cwd(), 'src', 'data', 'articles', 'blog');