+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.5.0] - Unreleased
 
+### Added
+- canonical URLs to website pages ([241](https://github.com/boarlabsxyz/Kavoon/pull/241))
+
 ### Fixed
 - minor vulnerabilities from 16 to 6 ([240](https://github.com/boarlabsxyz/Kavoon/pull/240))
 - warnings in terminal when run `npm install` ([240](https://github.com/boarlabsxyz/Kavoon/pull/240))